+++ mod-source-record-manager-server/src/main/java/org/folio/services/mappers/processor/parameters/MappingParametersBuilder.java
@@ -8,6 +8,8 @@ import org.folio.dataimport.util.OkapiConnectionParams;
 import org.folio.dataimport.util.RestUtil;
 import org.folio.rest.jaxrs.model.ClassificationType;
 import org.folio.rest.jaxrs.model.ClassificationTypes;
+import org.folio.rest.jaxrs.model.ContributorNameType;
+import org.folio.rest.jaxrs.model.ContributorNameTypes;
 import org.folio.rest.jaxrs.model.IdentifierType;
 import org.folio.rest.jaxrs.model.IdentifierTypes;
 import org.folio.rest.jaxrs.model.InstanceType;
@@ -26,10 +28,12 @@ public class MappingParametersBuilder {
   private static final String CLASSIFICATION_TYPES_URL = "/classification-types";
   private static final String INSTANCE_TYPES_URL = "/instance-types";
   private static final String INSTANCE_FORMATS_URL = "/instance-formats";
+  private static final String CONTRIBUTOR_NAME_TYPES_URL = "/contributor-name-types";
   private static final String IDENTIFIER_TYPES_RESPONSE_PARAM = "identifierTypes";
   private static final String CLASSIFICATION_TYPES_RESPONSE_PARAM = "classificationTypes";
   private static final String INSTANCE_TYPES_RESPONSE_PARAM = "instanceTypes";
   private static final String INSTANCE_FORMATS_RESPONSE_PARAM = "instanceFormats";
+  private static final String CONTRIBUTOR_NAME_TYPES_RESPONSE_PARAM = "contributorNameTypes";
 
   private MappingParametersBuilder() {
   }
@@ -39,13 +43,16 @@ public class MappingParametersBuilder {
     Future<List<ClassificationType>> classificationTypesFuture = getClassificationTypes(params);
     Future<List<InstanceType>> instanceTypesFuture = getInstanceTypes(params);
     Future<List<InstanceFormat>> instanceFormatsFuture = getInstanceFormats(params);
-    return CompositeFuture.all(identifierTypesFuture, classificationTypesFuture, instanceTypesFuture, instanceFormatsFuture)
+    Future<List<ContributorNameType>> contributorNameTypesFuture = getContributorNameTypes(params);
+    return CompositeFuture.all(identifierTypesFuture, classificationTypesFuture, instanceTypesFuture,
+      instanceFormatsFuture, contributorNameTypesFuture)
       .map(ar ->
         new MappingParameters()
           .withIdentifierTypes(identifierTypesFuture.result())
           .withClassificationTypes(classificationTypesFuture.result())
           .withInstanceTypes(instanceTypesFuture.result())
           .withInstanceFormats(instanceFormatsFuture.result())
+          .withContributorNameTypes(contributorNameTypesFuture.result())
       );
   }
 
@@ -129,6 +136,28 @@ public class MappingParametersBuilder {
         if (response != null && response.containsKey(INSTANCE_FORMATS_RESPONSE_PARAM)) {
           List<InstanceFormat> instanceFormatList = response.mapTo(InstanceFormats.class).getInstanceFormats();
           future.complete(instanceFormatList);
+        } else {
+          future.complete(Collections.emptyList());
+        }
+      }
+    });
+    return future;
+  }
+
+  /**
+   * Requests for Contributor name types from application Settings (mod-inventory-storage)
+   *
+   * @param params Okapi connection parameters
+   * @return List of Contributor name types
+   */
+  private static Future<List<ContributorNameType>> getContributorNameTypes(OkapiConnectionParams params) {
+    Future<List<ContributorNameType>> future = Future.future();
+    RestUtil.doRequest(params, CONTRIBUTOR_NAME_TYPES_URL, HttpMethod.GET, null).setHandler(ar -> {
+      if (RestUtil.validateAsyncResult(ar, future)) {
+        JsonObject response = ar.result().getJson();
+        if (response != null && response.containsKey(CONTRIBUTOR_NAME_TYPES_RESPONSE_PARAM)) {
+          List<ContributorNameType> contributorNameTypes = response.mapTo(ContributorNameTypes.class).getContributorNameTypes();
+          future.complete(contributorNameTypes);
         } else {
           future.complete(Collections.emptyList());
         }