+++ NEWS.md
@@ -2,6 +2,7 @@
  * Added instance-type settings loading for mapping
  * Added classification settings loading for mapping
  * Added instance-format settings loading for mapping
+ * Added contributor name types setting loading for mapping
 
 ## 2019-09-09 v1.6.0
  * Progress mechanism was updated